Fix trailing <br>s added in lists with EXTENSION_HARD_LINE_BREAK enabled
Previously, with EXTENSION_HARD_LINE_BREAK enabled, <li>s would contain unnecessary <br />s at the end of each item. It was especially noticeable by the vertical space this left at the end of an ordered or unordered list.

This fixes that by stripping trailing <br />s from individual list items while in the process of stripping trailing newlines.

One shortcoming of this patch: it does not fix the issue in *nested* lists.